# Faster structure scores