### Your checklist for this pull request
Please review the [guidelines for contributing](CONTRIBUTING.md) to this repository.

- [x] Make sure you are requesting to **pull a topic/feature/bugfix branch** (right side). Don't request your master!
- [x] Make sure you are making a pull request against the **dev branch** (left side). Also you should start *your branch* off *our dev*.
- [x] Check the commit's or even all commits' message styles matches our requested structure.

### List of changes to the codebase in this pull request
- Use numpy/scipy operations to speed up the structure scores. Computing the score is 70-100x faster on 100k samples from Alarm.

#### Before (604f1f9)
```python
In [1]: import pandas as pd

In [2]: from pgmpy.utils import get_example_model

In [3]: from pgmpy.estimators import BicScore, K2Score, BDeuScore

In [4]: model = get_example_model('alarm')

In [5]: samples = pd.read_pickle('samples.pkl')

In [6]: samples.info()
<class 'pandas.core.frame.DataFrame'>
RangeIndex: 100000 entries, 0 to 99999
Data columns (total 37 columns):
 #   Column        Non-Null Count   Dtype
---  ------        --------------   -----
 0   MINVOLSET     100000 non-null  object
 1   VENTMACH      100000 non-null  object
 ...
 35  PCWP          100000 non-null  object
 36  CVP           100000 non-null  object
dtypes: object(37)
memory usage: 28.2+ MB

In [7]: k2 = K2Score(data=samples)

In [8]: k2.score(model)  # Burn-in to cache state counts
Out[8]: -1048085.900814158

In [9]: %timeit k2.score(model)
263 ms ± 67.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [10]: bdeu = BDeuScore(data=samples)

In [11]: bdeu.score(model)  # Burn-in to cache state counts
Out[11]: -1047837.8490997906

In [12]: %timeit bdeu.score(model)
199 ms ± 5.89 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [13]: bic = BicScore(data=samples)

In [14]: bic.score(model)  # Burn-in to cache state counts
Out[14]: -1048844.320202879

In [15]: %timeit bic.score(model)
247 ms ± 5.31 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

#### After
```python
In [1]: import pandas as pd

In [2]: from pgmpy.utils import get_example_model

In [3]: from pgmpy.estimators import BicScore, K2Score, BDeuScore

In [4]: model = get_example_model('alarm')

In [5]: samples = pd.read_pickle('samples.pkl')

In [6]: samples.info()
<class 'pandas.core.frame.DataFrame'>
RangeIndex: 100000 entries, 0 to 99999
Data columns (total 37 columns):
 #   Column        Non-Null Count   Dtype
---  ------        --------------   -----
 0   MINVOLSET     100000 non-null  object
 1   VENTMACH      100000 non-null  object
 ...
 35  PCWP          100000 non-null  object
 36  CVP           100000 non-null  object
dtypes: object(37)
memory usage: 28.2+ MB

In [7]: k2 = K2Score(data=samples)

In [8]: k2.score(model)  # Burn-in to cache state counts
Out[8]: -1048085.900814158

In [9]: %timeit k2.score(model)
2.65 ms ± 20.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [10]: bdeu = BDeuScore(data=samples)

In [11]: bdeu.score(model)  # Burn-in to cache state counts
Out[11]: -1047837.8490997914

In [12]: %timeit bdeu.score(model)
2.74 ms ± 37.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [13]: bic = BicScore(data=samples)

In [14]: bic.score(model)  # Burn-in to cache state counts
Out[14]: -1048844.3202028787

In [15]: %timeit bic.score(model)
2.94 ms ± 34.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```